Refactor: Add numbering to higher-level geom module files
