+++ automol/geom/_0molsym.py
